--- tests/sample.yaml
@@ -1,10 +1,0 @@
-tests:
-  kernel.threads.apis:
-    tags: kernel threads userspace ignore_faults
-    min_flash: 34
-  kernel.threads.apis.pinonly:
-    tags: kernel threads userspace ignore_faults
-    min_flash: 34
-    filter: CONFIG_SMP
-    extra_configs:
-      - CONFIG_SCHED_CPU_MASK_PIN_ONLY=y
