--- twister2/__init__.py
@@ -1,1 +1,0 @@
-from .test_loader import TwisterSpec, load_c_test
