+++ twister2/plugin.py
@@ -1,17 +1,25 @@
 import logging
+from pathlib import Path
 
 import pytest
+
+from twister2.yaml_file import YamlFile
 
 from .report.test_plan_plugin import TestPlanPlugin
 
 logger = logging.getLogger(__name__)
 
 
-@pytest.fixture(scope='module')
-def dut(request: pytest.FixtureRequest):
-    logger.info('Dut fixture...')
-    logger.info(request.config.args)
-    return dict(name='dut1')
+# @pytest.fixture(scope='module')
+# def dut(request: pytest.FixtureRequest):
+#     logger.info('Dut fixture...')
+#     logger.info(request.config.args)
+#     return dict(name='dut1')
+
+
+def pytest_collect_file(parent, path):
+    if path.basename in ("sample.yaml", 'testcase.yaml'):
+        return YamlFile.from_parent(parent, path=Path(path))
 
 
 def pytest_addoption(parser):
